create new helper for waiting for igv rendering
